requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.0
